test: utilize pytest
