+++ .gitignore
@@ -1,6 +1,10 @@
 # Created by https://www.toptal.com/developers/gitignore/api/python
 # Edit at https://www.toptal.com/developers/gitignore?templates=python
 
+*.sqlite
+*.sqlite-journal
+*/*.sqlite
+
 ### Python ###
 # Byte-compiled / optimized / DLL files
 __pycache__/
